с Sealedd примерно разобрался

+++ src/by/tms/lesson17/onl30/record_and_sealed/sealed/Apple.java
@@ -2,7 +2,7 @@ package by.tms.lesson17.onl30.record_and_sealed.sealed;
 
 public final class Apple extends Fruit {
 
-    public Apple(String name, int price) {
+    public Apple(String name, double price) {
         super(name, price);
     }
 }
