+++ src/by/tms/lesson17/onl30/record_and_sealed/sealed/Cherry.java
@@ -1,7 +1,7 @@
 package by.tms.lesson17.onl30.record_and_sealed.sealed;
 
 public non-sealed class Cherry extends Fruit{
-    public Cherry(String name, int price) {
+    public Cherry(String name, double price) {
         super(name, price);
     }
 }
