+++ src/by/tms/lesson17/onl30/record_and_sealed/sealed/Pear.java
@@ -1,7 +1,7 @@
 package by.tms.lesson17.onl30.record_and_sealed.sealed;
 
 public sealed class Pear extends Fruit permits SortPear {
-    public Pear(String name, int price) {
+    public Pear(String name, double price) {
         super(name, price);
     }
 }
