+++ src/by/tms/lesson17/onl30/record_and_sealed/sealed/SortPear.java
@@ -1,7 +1,7 @@
 package by.tms.lesson17.onl30.record_and_sealed.sealed;
 
 public final class SortPear extends Pear {
-    public SortPear(String name, int price) {
+    public SortPear(String name, double price) {
         super(name, price);
     }
 }
